SK-2071 update error handling, message and validation of detect text with tests

+++ src/main/java/com/skyflow/vault/detect/EntityInfo.java
@@ -9,9 +9,9 @@ public class EntityInfo {
     private final TextIndex textIndex;
     private final TextIndex processedIndex;
     private final String entity;
-    private final Map<String, Float> scores;
+    private final Map<String, Double> scores;
 
-    public EntityInfo(String token, String value, TextIndex textIndex, TextIndex processedIndex, String entity, java.util.Map<String, Float> scores) {
+    public EntityInfo(String token, String value, TextIndex textIndex, TextIndex processedIndex, String entity, java.util.Map<String, Double> scores) {
         this.token = token;
         this.value = value;
         this.textIndex = textIndex;
@@ -40,7 +40,7 @@ public class EntityInfo {
         return entity;
     }
 
-    public Map<String, Float> getScores() {
+    public Map<String, Double> getScores() {
         return scores;
     }
 }
